+++ .gitignore
@@ -14,3 +14,5 @@ target
 omod/target/
 api/target/
 *.iml
+
+.vscode
