+++ README.md
@@ -1,8 +1,12 @@
-# Patient Identifier Generator
+# OpenMRS Module LIME EMR
 
-This repository helps in generating patient identifier dynamically for MSF-OCG
+This module contains customizations made to the OpenMRS platform for the LIME EMR project as well as other features that are not specific to the LIME EMR project.
+
+## Features
+
+### Patient Identifier Generator
+This is responsible for overriding OpenMRS ID as the default and replacing it with the MSF ID. The MSF ID is generated by the system and like the OpenMRS ID is unique to each patient. The MSF ID is generated by concatenating the prefix `IQ146-` with a unique number. The unique number is generated by incrementing the last number in the database by 1 at patient creation.
 
 With this module we are allowing the user to have an Id with an hyphen (`-`) after every group of three digits in the Identifier.
 
-<b>Prefix:</b> This is set to `IQ146-`
-<br>
+### Metadata Sharing
