+++ api/pom.xml
@@ -3,14 +3,14 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.openmrs.module</groupId>
-		<artifactId>patientidentifiergenerator</artifactId>
+		<artifactId>lime-emr</artifactId>
 		<version>0.1.0</version>
 	</parent>
 
-	<artifactId>patientidentifiergenerator-api</artifactId>
+	<artifactId>lime-emr-api</artifactId>
 	<packaging>jar</packaging>
-	<name>Patient Identifier Generator API</name>
-	<description>API project for Patient Identifier Generator</description>
+	<name>LIME EMR API</name>
+	<description>API project for LIME EMR dependencies and customizations</description>
 
 	<properties>
 		<openMRSVersion>2.6.6</openMRSVersion>
@@ -100,5 +100,5 @@
 			</plugin>
 		</plugins>
 	</build>
-	
+
 </project>
