+++ api/src/main/java/org/openmrs/module/lime/MosulMetadataUserListener.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.patientidentifiergenerator;
+package org.openmrs.module.lime;
 
 import org.openmrs.Role;
 import org.openmrs.User;
