+++ api/src/main/resources/moduleApplicationContext.xml
@@ -10,6 +10,6 @@
   		    http://www.springframework.org/schema/context
   		    http://www.springframework.org/schema/context/spring-context-3.0.xsd">
 
-    <context:component-scan base-package="org.openmrs.module.patientidentifiergenerator" />
-    
+    <context:component-scan base-package="org.openmrs.module.lime" />
+
 </beans>
