+++ omod/pom.xml
@@ -5,14 +5,14 @@
 
 	<parent>
 		<groupId>org.openmrs.module</groupId>
-		<artifactId>patientidentifiergenerator</artifactId>
+		<artifactId>lime-emr</artifactId>
 		<version>0.1.0</version>
 	</parent>
 
-	<artifactId>patientidentifiergenerator-omod</artifactId>
+	<artifactId>lime-emr-omod</artifactId>
 	<packaging>jar</packaging>
-	<name>MSF-OCG Patient Identifier Generator Module OMOD</name>
-	<description>OMOD project for Patient Identifier Generator</description>
+	<name>LIME EMR OMOD</name>
+	<description>OMOD project for LIME EMR dependencies and customizations</description>
 
 	<properties>
 		<openMRSVersion>2.6.6</openMRSVersion>
