+++ omod/src/main/resources/config.xml
@@ -2,20 +2,20 @@
 
 <module configVersion="1.2">
 
-	<id>patientidentifiergenerator</id>
-	<name>MSF-OCG Patient Identifier Generator Module</name>
+	<id>lime-emr</id>
+	<name>MSF-OCG LIME EMR Module</name>
 	<version>0.1.0</version>
-	<package>org.openmrs.module.patientidentifiergenerator</package>
+	<package>org.openmrs.module.lime</package>
 	<author>msf</author>
 	<description>
-		Helps to generate Patient Identifier
+		Base module for LIME backend dependencies and customizations like the Patient Identifier
 	</description>
 
-	<activator>org.openmrs.module.patientidentifiergenerator.PatientIdentifierGeneratorActivator</activator>
+	<activator>org.openmrs.module.lime.LimeEmrActivator</activator>
 
 	<advice>
 		<point>org.openmrs.api.PatientService</point>
-		<class>org.openmrs.module.patientidentifiergenerator.advice.BeforeSaveAdvice</class>
+		<class>org.openmrs.module.lime.advice.BeforeSaveAdvice</class>
 	</advice>
 
 	<require_modules>
@@ -23,5 +23,5 @@
 		<require_module version="2.2.0">org.openmrs.module.datafilter</require_module>
 		<require_module version="1.0.0">org.bahmni.module.appointments</require_module>
 	</require_modules>
-	
+
 </module>
